IllegalArgumentException is better than DbException in case of bug #3.
If the app restarts after crashing and tries to return to the activity
where it crashed, DatabaseConfig's encryption key will be null and no DB
operations can be performed. It's better to crash again in that case
than for all DB operations to fail.

Fixing bug #3 will make this irrelevant.